@@ -1,51 +1,6 @@
-import * as esbuild from "esbuild-wasm";
-import ReactDOM from "react-dom";
-import { useState, useEffect, useRef } from "react";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App';
 
-function App() {
-  const ref = useRef<any>();
-  const [input, setInput] = useState("");
-  const [code, setCode] = useState("");
-
-  const startService = async () => {
-    await esbuild.initialize({
-      worker: true,
-      wasmURL: "/esbuild.wasm",
-    });
-    ref.current = true;
-  };
-
-  useEffect(() => {
-    startService();
-  }, []);
-
-  const onClick = () => {
-    if (!ref.current) {
-      return;
-    }
-
-    esbuild
-      .transform(input, {
-        loader: "jsx",
-        target: "es2015",
-      })
-      .then((result) => {
-        setCode(result.code);
-      });
-  };
-
-  return (
-    <>
-      <textarea
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-      ></textarea>
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <pre>{code}</pre>
-    </>
-  );
-}
-
-ReactDOM.render(<App />, document.querySelector("#root"));
+const root = ReactDOM.createRoot(document.getElementById('root')!);
+root.render(<App />);
